# backend/swagger.yaml
swagger: "2.0"
info:
  title: ProFilm eWarranty API
  description: API for managing ProFilm warranty claims, products, shops, and more
  version: "1.0.0"
  contact:
    name: ProFilm Support
    email: support@profilm.com
host: localhost:8080
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
...
paths:
...
  /products/brands:
    post:
      tags:
        - Product Brands
      summary: Create a new product brand
      description: Create a new product brand in the system
      parameters:
        - in: body
          name: brand
          description: Product brand object
          required: true
          schema:
            $ref: "#/definitions/CreateProductBrandRequest"
      responses:
        201:
          description: Product brand created successfully
          schema:
            $ref: "#/definitions/ProductBrandResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Product Brands
      summary: List all product brands
      description: Retrieve a list of all product brands
      responses:
        200:
          description: List of product brands
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductBrandResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products/types:
    post:
      tags:
        - Product Types
      summary: Create a new product type
      description: Create a new product type under a brand
      parameters:
        - in: body
          name: type
          description: Product type object
          required: true
          schema:
            $ref: "#/definitions/CreateProductTypeRequest"
      responses:
        201:
          description: Product type created successfully
          schema:
            $ref: "#/definitions/ProductTypeResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Product Types
      summary: List all product types
      description: Retrieve a list of all product types
      responses:
        200:
          description: List of product types
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductTypeResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products/series:
    post:
      tags:
        - Product Series
      summary: Create a new product series
      description: Create a new product series under a product type
      parameters:
        - in: body
          name: series
          description: Product series object
          required: true
          schema:
            $ref: "#/definitions/CreateProductSeriesRequest"
      responses:
        201:
          description: Product series created successfully
          schema:
            $ref: "#/definitions/ProductSeriesResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Product Series
      summary: List all product series
      description: Retrieve a list of all product series
      responses:
        200:
          description: List of product series
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductSeriesResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products/names:
    post:
      tags:
        - Product Names
      summary: Create a new product name
      description: Create a new product name under a product series
      parameters:
        - in: body
          name: productName
          description: Product name object
          required: true
          schema:
            $ref: "#/definitions/CreateProductNameRequest"
      responses:
        201:
          description: Product name created successfully
          schema:
            $ref: "#/definitions/ProductNameResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Product Names
      summary: List all product names
      description: Retrieve a list of all product names
      responses:
        200:
          description: List of product names
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductNameResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products/warranty-years:
    post:
      tags:
        - Warranty Years
      summary: Create a new warranty year
      description: Create a new warranty year option
      parameters:
        - in: body
          name: warrantyYear
          description: Warranty year object
          required: true
          schema:
            $ref: "#/definitions/CreateWarrantyYearRequest"
      responses:
        201:
          description: Warranty year created successfully
          schema:
            $ref: "#/definitions/WarrantyYearResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Warranty Years
      summary: List all warranty years
      description: Retrieve a list of all warranty year options
      responses:
        200:
          description: List of warranty years
          schema:
            type: array
            items:
              $ref: "#/definitions/WarrantyYearResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products:
    post:
      tags:
        - Products
      summary: Create a new product
      description: Create a new product in the system
      parameters:
        - in: body
          name: product
          description: Product object
          required: true
          schema:
            $ref: "#/definitions/CreateProductRequest"
      responses:
        201:
          description: Product created successfully
          schema:
            $ref: "#/definitions/ProductResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Products
      summary: List all products
      description: Retrieve a list of all products
      responses:
        200:
          description: List of products
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product ID
      responses:
        200:
          description: Product details
          schema:
            $ref: "#/definitions/ProductResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Products
      summary: Update product
      description: Update an existing product
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product ID
        - in: body
          name: product
          description: Updated product object
          required: true
          schema:
            $ref: "#/definitions/UpdateProductRequest"
      responses:
        200:
          description: Product updated successfully
          schema:
            $ref: "#/definitions/ProductResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product from the system
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product ID
      responses:
        204:
          description: Product deleted successfully
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Product Detail Routes
  /products/details:
    get:
      tags:
        - Products
      summary: List all products with details
      description: Retrieve a list of all products with comprehensive details
      responses:
        200:
          description: List of products with details
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductDetailResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /products/details/{id}:
    get:
      tags:
        - Products
      summary: Get product details by ID
      description: Retrieve comprehensive product details by ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product ID
      responses:
        200:
          description: Product details
          schema:
            $ref: "#/definitions/ProductDetailResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # States Endpoints
  /states:
    post:
      tags:
        - States
      summary: Create a new state
      description: Create a new state in the system
      parameters:
        - in: body
          name: state
          description: State object
          required: true
          schema:
            $ref: "#/definitions/CreateStateRequest"
      responses:
        201:
          description: State created successfully
          schema:
            $ref: "#/definitions/StateResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - States
      summary: List all states
      description: Retrieve a list of all states
      responses:
        200:
          description: List of states
          schema:
            type: array
            items:
              $ref: "#/definitions/StateResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /states/{id}:
    get:
      tags:
        - States
      summary: Get state by ID
      description: Retrieve a specific state by its ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: State ID
      responses:
        200:
          description: State details
          schema:
            $ref: "#/definitions/StateResponse"
        404:
          description: State not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - States
      summary: Update state
      description: Update an existing state
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: State ID
        - in: body
          name: state
          description: Updated state object
          required: true
          schema:
            $ref: "#/definitions/UpdateStateRequest"
      responses:
        200:
          description: State updated successfully
          schema:
            $ref: "#/definitions/StateResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: State not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - States
      summary: Delete state
      description: Delete a state from the system
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: State ID
      responses:
        204:
          description: State deleted successfully
        404:
          description: State not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Shop Endpoints
  /shops:
    post:
      tags:
        - Shops
      summary: Create a new shop
      description: Create a new shop in the system
      parameters:
        - in: body
          name: shop
          description: Shop object
          required: true
          schema:
            $ref: "#/definitions/CreateShopRequest"
      responses:
        201:
          description: Shop created successfully
          schema:
            $ref: "#/definitions/ShopResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Shops
      summary: List all shops
      description: Retrieve a list of all shops
      responses:
        200:
          description: List of shops
          schema:
            type: array
            items:
              $ref: "#/definitions/ShopResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /shops/{id}:
    get:
      tags:
        - Shops
      summary: Get shop by ID
      description: Retrieve a specific shop by its ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Shop ID
      responses:
        200:
          description: Shop details
          schema:
            $ref: "#/definitions/ShopResponse"
        404:
          description: Shop not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Shops
      summary: Update shop
      description: Update an existing shop
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Shop ID
        - in: body
          name: shop
          description: Updated shop object
          required: true
          schema:
            $ref: "#/definitions/UpdateShopRequest"
      responses:
        200:
          description: Shop updated successfully
          schema:
            $ref: "#/definitions/ShopResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Shop not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Shops
      summary: Delete shop
      description: Delete a shop from the system
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Shop ID
      responses:
        204:
          description: Shop deleted successfully
        404:
          description: Shop not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Shop Detail Routes
  /shops/details:
    get:
      tags:
        - Shops
      summary: List all shop details
      description: Retrieve a list of all shops with comprehensive details
      responses:
        200:
          description: List of shops with details
          schema:
            type: array
            items:
              $ref: "#/definitions/ShopDetailResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /shops/details/{id}:
    get:
      tags:
        - Shops
      summary: Get shop details by ID
      description: Retrieve comprehensive shop details by ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Shop ID
      responses:
        200:
          description: Shop details
          schema:
            $ref: "#/definitions/ShopDetailResponse"
        404:
          description: Shop not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Product Allocation Endpoints
  /allocations:
    post:
      tags:
        - Product Allocations
      summary: Create a new product allocation
      description: Create a new product allocation to a shop
      parameters:
        - in: body
          name: allocation
          description: Product allocation object
          required: true
          schema:
            $ref: "#/definitions/CreateProductAllocationRequest"
      responses:
        201:
          description: Product allocation created successfully
          schema:
            $ref: "#/definitions/ProductAllocationResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Product Allocations
      summary: List all product allocations
      description: Retrieve a list of all product allocations
      responses:
        200:
          description: List of product allocations
          schema:
            type: array
            items:
              $ref: "#/definitions/ProductAllocationResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /allocations/{id}:
    get:
      tags:
        - Product Allocations
      summary: Get product allocation by ID
      description: Retrieve a specific product allocation by its ID
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product allocation ID
      responses:
        200:
          description: Product allocation details
          schema:
            $ref: "#/definitions/ProductAllocationResponse"
        404:
          description: Product allocation not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Product Allocations
      summary: Update product allocation
      description: Update an existing product allocation
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product allocation ID
        - in: body
          name: allocation
          description: Updated product allocation object
          required: true
          schema:
            $ref: "#/definitions/UpdateProductAllocationRequest"
      responses:
        200:
          description: Product allocation updated successfully
          schema:
            $ref: "#/definitions/ProductAllocationResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Product allocation not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Product Allocations
      summary: Delete product allocation
      description: Delete a product allocation from the system
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Product allocation ID
      responses:
        204:
          description: Product allocation deleted successfully
        404:
          description: Product allocation not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Car Parts Endpoints
  /car-parts:
    post:
      tags:
        - Car Parts
      summary: Create a new car part
      description: Create a new car part in the system
      parameters:
        - in: body
          name: carPart
          description: Car part object
          required: true
          schema:
            $ref: "#/definitions/CreateCarPartRequest"
      responses:
        201:
          description: Car part created successfully
          schema:
            $ref: "#/definitions/CarPartResponse"
        400:
          description: Invalid request payload
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Car Parts
      summary: List all car parts
      description: Retrieve a list of all car parts
      responses:
        200:
          description: List of car parts
          schema:
            type: array
            items:
              $ref: "#/definitions/CarPartResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        description: Error message
      code:
        type: integer
        description: Error code
    required:
      - error

  # Product Brand Models
  CreateProductBrandRequest:
    type: object
    properties:
      name:
        type: string
        description: Brand name
        example: "Profilm"
    required:
      - name
...
  ProductBrandResponse:
    type: object
    properties:
      id:
        type: string
        format: uuid
        description: Brand ID
      name:
        type: string
        description: Brand name
      isActive:
        type: boolean
        description: Whether the brand is active
      createdAt:
        type: string
        format: date-time
        description: Creation timestamp
      updatedAt:
        type: string
        format: date-time
        description: Last update timestamp

  # Product Type Models
  CreateProductTypeRequest:
    type: object
    properties:
      brandId:
        type: string
        format: uuid
        description: Brand ID this type belongs to
      name:
        type: string
        description: Type name
        example: "Window Tinting Film"
      description:
        type: string
        description: Type description
    required:
      - brandId
      - name
...
  ProductTypeResponse:
    type: object
    properties:
      id:
        type: string
        format: uuid
        description: Type ID
      brandId:
        type: string
        format: uuid
        description: Brand ID
      name:
        type: string
        description: Type name
      description:
        type: string
        description: Type description
      isActive:
        type: boolean
        description: Whether the type is active
      createdAt:
        type: string
        format: date-time
        description: Creation timestamp
      updatedAt:
        type: string
        format: date-time
        description: Last update timestamp

  # Product Models
...
  CreateWarrantyYearRequest:
    type: object
    properties:
      years:
        type: integer
        description: Number of warranty years
        example: 5
    required:
      - years
...
  UpdateStateRequest:
    type: object
    properties:
      name:
        type: string
        description: Updated state name
...